move accelerometer library into a separate file
